# swagger/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/get/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "取得用戶",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用戶ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/main.User"
                        }
                    },
                    "400": {
                        "description": "參數錯誤",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "內部錯誤",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/update_user": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "更新用戶",
                "parameters": [
                    {
                        "description": "用戶信息",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "參數錯誤",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "內部錯誤",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.User": {
            "type": "object",
            "properties": {
                "age": {
                    "description": "年齡",
                    "type": "integer"
                },
                "id": {
                    "description": "用戶ID",
                    "type": "integer"
                },
                "name": {
                    "description": "姓名",
                    "type": "string"
                }
            }
        }
    }
}